Merged PR 8071: Performance improvement test
This change is to try out if we can prevent SQL from doing a full table scan on the clustered key when there is a covering index.

The idea is that there will be no cost for lookups going from the index to fetch all the fields of the record.

Related work items: #4758