PR 8305: Obsoleting POS keyboard binding module
The keybinding module was extremely slow, and was locking across all other sessions when opening the POS page (meaning if multiple machines open the POS, in the morning or after a NST crash etc.).
Based on old cases it was originally developed because the escape button could cancel POS sales in classic 2009 POS, but other than that it has seen almost zero usage since it was introduced.
It was decided to obsolete the module.

Case 480383
Related work items: #4872