PR 8310: Field "Blocking UI" removed from POS Menu Button table
The field didn't really make sense as it should not be persistent data stored on a per button level. Either an action is intended to have async UI or it is not. No consultant or user will ever know when to check or uncheck this checkmark when using the system, it should be hardcoded by the developer that created the action.

Case 482901
Related work items: #4873