Merged PR 8571: CalcDate without a reference date on "NP Retail Manager" profile page
The issue is related to this error:
![image (2).png](https://dev.azure.com/navipartner/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/8571/attachments/image%20%282%29.png)

The error started to appear for different customers just recently.
Unfortunately I couldn't reproduce the error on a container in order to be able to debug and test the solution, because the error seems to be raised only on production server. This must be either platform, or server configuration specific.
Since debugging cannot be enabled for tenants running on production servers, I just tried to find the most probable reason. We'll have to publish this solution to a customer database having the issue, and see, if it helped.

So the error seems to be generated in page 6151260 "NPR POS Entry Cue".
There is only one piece of code dealing with dates:
`Rec.SetRange("EFT Errors Date Filter", CalcDate('<-30D>'), Today());`
It is in trigger OnOpenPage(), and the line was added just recently - 2 months ago in this PR: !7201
In this case the CalcDate function does not have a reference date. Although this is an optional parameter (the current system date should be used, if this is omitted), system might not be able to get current system date and uses undefined date/time instead. Could it be a platform bug by Microsoft?
I have added Today() as implicit reference date. Let's see if it solves the issue. Otherwise the new variant of the code should act exactly the same as the old one.
Case 482711

Related work items: #4951