Merged PR 8665: SMS Setup missing on first company open
SMS Setup missing on first company open. For example, when you go to companies, create new one and make change to it. After automatic logout and login to that company, "something went wrong message" is being displayed.

Because of InitCompany it ends up in codeunit 6059940 "NPR SMS Management", procedure GetJobQueueCategoryCode() where it tries SMSSetup.Get() but it will be empty on first open.
I have changed it to initialize by existing function from another codeunit, but maybe this is not the smartest solution. I have added @Filip and @Tin because they are more familiar with this functionality.

Related work items: #4981