Merged PR 8749: Add new logic for generating barcode and barcode size
On BC18 for printing the shelf label report, we use fonts instead of generating images. Because of that, we need to make sure that the whole barcode would fit into the textbox.
As agreed on the call with Mikkel, we are generating barcodes for the maximum length of 20 characters. If a barcode has more than 20 characters then the code will display ***.

Related work items: #4991