PR 8772: Retail Manager RC: page background tasks
Moved flow field pre-calculation on retail manager role centre to background tasks. This should improve the performance of the page, as the computations are going to be run now asynchronously in background. As a result the page should become quicker to open and more responsive, faster for users to enter information. Users won't be blocked from working while waiting for the computations to finish.

Case 485974
Related work items: #5012