Merged PR 8836: Refactor Replication module to use generic BC API import
Refactored module to be easier to extend and setup.

- Created one Codeunit that can handle import for any BC table instead of one Codeunit handler for each table how it was previously.
- Created table "Replication Special Fields Mapping" where we define mapping of fields for each endpoint. Predefined mappings are automatically generated when endpoints are registered (using discovery event pattern).
- In case no mapping exist, system tries to find field in the API response by converting the table field name to camel case. So only exceptions need to be mapped.

Introduced also other new features that make module more configurable, examples: possibility to configure if OnValidate trigger should be run for any field, possibility to Skip a field from the import, possibility to configure if OnINsert/OnModify triggers are run when a record is created/modified etc.

Images (Media/MediaSet) are handled for Customer, Item and Vendor.

Now this Codeunit can also handle fields from customer extension. So in that case we only need to create a copy of the API page in the customer extension and Special Fields Mappings if the case.

Related work items: #5043