Merged PR 9118: Fix NaviConnect only processing one task
Changing the values of a variable that's used in an iterator is not good. This shows here where the `SetCurrentKey()` makes it so only one task is processed.

This PR rectifies this issue for NaviConnect Task processing.

(Also bonus fix in that "Show Processed" on Task List would only show Processed tasks. It should show all tasks).

Related work items: #5108