Merged PR 9197: Set max sync time for Job Queue codeunit to run
Since a codeunit currently being run on the Task Scheduler cannot be canceled, it's nice with a way to intervene if a sync goes wild. This PR introduces a max limit of 30 minutes of runtime while processing tasks. The same limit existed when this codeunit ran with Task Queue.

Related work items: #5115