Merged PR 9221: Case 490323. Processing multiple import types with a Job Queue Entry
Added possibility to process multiple import types with a single Job Queue Entry.
We can now specify a filter for import type codes as a parameter for Job Queue Entry.
This is needed to support Shopify integration.
I have also refactored the way Job Queue parameters are manipulated for import/task list processing jobs, in order to have unified approach. It can be extended later on (if needed) with events, to which extensions can subscribe to.
Case 490323

Related work items: #5130