PR 9411: Increased Maximum No. of Attempts for Retention Policy Job Queues
This is because of an alleged bug in MS solution, when some of the retention policy run attempts may finish with an error "Arithmetic operation resulted in an overflow". We've found out that repeating the attempts eventually results in a successful run.
An integer variable storing the number of deleted records seemingly is calculated incorrectly rapidly rising to a very big value and eventually hitting the integer variable limit. But system still deletes some records on each attempt. So repeating multiple times should do the job.
This is a temporary workaround until we'll have a fix from Microsoft, of find another way around.

Case 490992

Related work items: #5172