Merged PR 9286: Optimize item search from ean box
Optimize item search from ean box

Guess is that most of the searches will be missed when doing the search description. Because of that we create 2 searches, first is with search description only, and if we found something, after that we perform additional search including a blocked filter.

Related work items: #5141