PR 9584: Printing via report/codeunit for retail vouchers
Previously only printing via print templates was supported for vouchers.
Now it is possible to select a report object or a codeunit for voucher printing.
The new fields were added to Voucher Type card and Voucher card.

The functionality was requested by Buster for customer Hartmanns.

Case 492244
Related work items: #5195