PR 9682: Restrict some general and vat business posting group usage on POS
When the "Default POS Posting Setup" is set to value "Customer" on POS Posting Profile, there still can be scenarios, when some business posting group usage may be restricted on POS.
Primarily this is related to cases, when a foreign customer is selected in a POS sale. Previously this could lead to a situation, when export transaction oriented VAT calculation setup (0% VAT) would be applied to the POS sale.
Now we can mark those General and VAT business groups as restricted on POS, and now system will use the groups defined for the POS store, instead of copying such posting groups from selected customer card.

Case 492448
Related work items: #5217