Merged PR 9944: Add custom action for creating a new customer from the worksheet page for smart search
Customer list page that users run from POS is the worksheet page. It is created that way because worksheet pages are always started in editable mode by default. We wanted to achieve that because of smart search.
One of the downsides of that approach is that worksheet pages don't have default action for creating a new record and opening a specific card page as we have on list pages and because of that in this PR I created a custom action for creating a new customer from worksheet page.

Related work items: #5264