Merged PR 10142: Adding NPR prefix to enum extension values
Couple things to note on this PR (and correct me if I'm wrong)
- Enum cannot be obsoleted
- Adding NPR to enum value isn't breaking change (if I'm only thinking of NPRetail app) if you don't change enum value ID
- If you change enum value ID you then "loose" (data is still in db as integer but it is not visible since there is no enum attached to the integer value) the data and you need to do conversion.

Since I'm not able to obsolete enum, i've just changed ID and added prefix. Code is changed and it is publishing to container.
Data is changed with upgrade codeunit using trick to filter on old integer value of enum and changing the value to new enum value.

I tested the upgrade on new container
1) create new container with np retail
2) populate some item references with reference type = Retail Serial No.
3) publish new version
4) check Item Reference type

Related work items: #5386