Merged PR 10178: Set specific newtonsoft version and reintroduce lightweight assemblyresolver for debugging.
Assemblyresolver is not hooked up to any npdeploy, it most likely won't be necessary if the explicit newtonsoft version is enough to fix this.
Assemblies can be uploaded to it manually from the page.

The reason v10 is chosen is because it appears to be what NST has been serving up (MS includes v10 in BC) so if we were able to be in prod for the last year-ish on an .app with zero hard version requirement, we might as well jump to the highest one that appears to work which is v10.

Related work items: #5404