Merged PR 10250: Replacing copy method for BC 18 and later on item smart search
The smart search for Items on BC18 and BC19 not working. The page just keeps hanging while trying to show new results. Investigating the problem we narrow it down that the Copy function in combination with Mark causing problems. This is weird because the same approach is working fine on BC17.
Even weirder is that we cannot reproduce the same behavior for smart search on Customers pages and Membership module in new versions of BC.
Since the workaround solution is a bit slower than the initial one, we decided to just add this workaround for the Items page. If a similar problem occurs for Customers or Membership module pages the same approach will be needed there also.

Related work items: #5405