Merged PR 10316: New End Of Day UX
New End Of Day UX

The new UX for end of day is still work in progress, but it is in state where it can be reviewed.
It will not be activated unless you specify that new UX should be used. The setup is on the **POS End of Day Profile** and the field **User Experience** is currently not visible (you will have to add it via the designer).

Related work items: #5441