Merged PR 10350: Ticket Valid From/Until follows duration formula
The field **Valid Until Date** on **Ticket** follows the **Duration Formula** specified on either **Ticket Type** or **Ticket BOM**.

When a ticket includes multiple admissions, the biggest **Duration Formula** is applied to Today() to calculate the **Valid Until Date**

The number of schedules selectable in POS and presented in the API (when an item is specified) is now determined by the **Duration Formula** also.

There is a small glitch when setting up a multi-admission ticket with different duration formulas, the initial duration is selected individually, but when you change the ticket, the maximum duration formula is applied.

Related work items: #5458