Merged PR 10636: Update Item SEO Link logic
In !7657 the SEO link logic was updated to include brand name due to a request from a customer. This changed the logic for all customers and not everybody wants it to be done this way.

So this PR is a mix between reverting that and adding extensibility. PTE apps can add the text they want to be part of the SEO link by subscribing to the `OnBeforeItemSeoLinkFormat()` event. They should then be calling `"Magento Item Mgt.".UpdateItemSeoLink()` in case they rely on other fields to be part of the SEO link. An example:

```al
codeunit 50100 "PTE Magento Subscriber"
{
    [EventSubscriber(ObjectType::Codeunit, Codeunit::"NPR Magento Item Mgt.", 'OnBeforeItemSeoLinkFormat', '', true, true)]
    local procedure UpdateSeoLink(Item: Record Item; var SeoLinkText: Text)
    begin
        SeoLinkText := Item."NPR Magento Name";
        if Item."NPR Magento Brand" <> '' then
            SeoLinkText := Item."NPR Magento Brand" + ' ' + SeoLinkText
    end;

    [EventSubscriber(ObjectType::Table, Database::Item, 'OnAfterValidateEvent', 'NPR Magento Brand', true, true)]
    local procedure OnAfterValidateMagentoBrand(var Rec: Record Item)
    var
        MagentoItemMgt: Codeunit "NPR Magento Item Mgt.";
    begin
        MagentoItemMgt.UpdateItemSeoLink(Rec);
    end;
}
```

Can be tested in container attached to case 501186. But I have of course also tested it myself.

Related work items: #5537