Merged PR 10861: Exit conditionally when all lines are not ready
Exit conditionally when all lines are not ready

When user does not fully complete end of day, some lines for counting are left in a "not ready" state. This is checked when pos entry is created. When not ready, no entry is created. This is now correctly handled so that posting is not attempted - which caused the error.

Related work items: #5619