Merged PR 10795: Commit on change view
This PR adds a commit to the general setview function.

The code after the commit is moved to a new function and the attribute CommitBehavior is added to assure there is no Commit when the change view request is sent to front end, and during the execution of data drivers collecting and pushing data to front end.

Also, a previously added Commit, when manually switching to login view was removed and replaced by this generic Commit.

Related work items: #5591