PR 10953: Lost sale resume process: allow leaving unfinished sale intact, if it belongs to another active session
This is to support multiple POS sessions, started for the same POS unit and under the same User ID (rare scenarios, but still possible).
Previously it was not possible because of lost/unfinished sale resume routine, which would result in any sale, started in an earlier POS session, being overtaken or cancelled by the next POS session.
Now, after user has initially confirmed resume or cancel of an existing sale, if system finds that the sale belongs to an active BC session, it will notify the user about it, and ask for additional confirmation.
If the user doesn't confirm that, system will skip the sale (leave the sale intact), and will go on with handling other unfinished sales, or with starting a new sale, if there are no other unfinished sales left.
There is one exception in regards to how unfinished POS sales with no lines (empty sales) are processed by the lost/unfinished sale resume routine: if the sale belongs to an active BC session, it will silently be skipped, otherwise it will be silently cancelled (previously any unfinished empty sale would be silently automatically cancelled).

Case 502780
Related work items: #5666