Merged PR 11012: MOCK NpXml Test. No more Magento, FTP and Azure downlad
MOCK NpXml Test unit so it doesn't do download from Azure Blob storage anymore and it's not using OttoSuenson Magento test server.

I was motivated to MOCK this in the way like Microsoft did for example with Email API. I even imported all their code and it was working like here:

![Test 1st with Standard.png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/11012/attachments/Test%201st%20with%20Standard.png) ![Test 1st with Standard 2.png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/11012/attachments/Test%201st%20with%20Standard%202.png)

And I was going step by step making the same thing in NpXml test but structure in our NpXML management is totally different and should be changed a lot. (Split it to additional codeunits, enums and interfaces which I didn't wanted to do because it is so connected with all other modules and by that it is sensitive to redesign. Plus it is working like a charm currently. Plus, there is no real reason for mocking Magento API call response because we are not handling it if it fail or not. No matter what the response is, we are only filling in the response blob, nothing more.)

FTP upload is completely removed because it can not be mocked. And besides that, the only difference in API and FTP is that XML is being uploaded to FTP instead posted to API.

One more thing, Codeunit 85003 "NPR Library - POS Mock" is changed because App couldn't be complied. One parameter was added in the meanwhile.

Related work items: #5635