PR 11073: Disallow running blocked POS actions
With this update we'll have means to block POS actions.
1. Added "Toggle Blocked" menu item on the page 6150703 "NPR POS Actions"
This will allow to block/unblock one or a range of POS actions.
System will actually run a new codeunit 6014691 "NPR Block/Unblock POS Action". Therefore we can limit permissions to run the functionality.
And there is an event called "OnBeforePosActionToggleBlocked" in the codeunit, to which we can subscribe to in cases, when we need to prevent users from blocking/unblocking certain POS actions (for example, if it is a compliance requirement in a country).
2. Adjusted POS action discovery process. Now the value of the "Blocked" field will be retained even after the POS action has been updated by the discovery process.
3. Introduced 2 means to prevent running blocked POS actions on POS:
   - system won't send to front-end POS menu buttons defined with blocked POS actions (just the same way as it is done now for the POS menu buttons themselves).
as a side effect, system won't send also now buttons defined for non-existing POS actions.
   - on attempt to run a blocked POS action system will show an error on the front-end
this is needed, because actions can be run on POS not only through buttons defined directly for POS actions, but also through named actions, or through nested workflows

Cases 503130, 503422.
Related work items: #5708, #5709