Merged PR 11097: Strange err. in Job Que mult. import types process
Strange error in Job Que multiple import types processing. And it's hard to catch because it doesn't show up always. In one month, only two times (from 21812 executions).

I know that at first look, my change with this "IF" statement isn't really needed here, because there is "IsEmpty" above it, but I have doubts that error from case 505364 and also from 506358 pops out because of this FindSet.

Codeunit which is processing import types is with us for a long time, but this function (ProcessImportEntries) was changed 3 months ago in this PR:

https://dev.azure.com/navipartner/NpCore/_git/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullrequest/9631?_a=files&path=/Application/src/NaviConnect/NcImportListProcessing.Codeunit.al

so it support processing multiple import types with a Job Queue Entry. In older version there was IF FindSet...

I will repeat one more time, there is no much sense that it's needed here, but I think it will be working solution. Or this one, or to add SelectLatestVersion(); before FindSet() because very frequent changes occur in the database (multiple import types, multiple jobs... repeating each 2 minutes and multiple other processes). So it can be that IsEmpty is more "Up to date" in question of database changes and return something but then FindSet doesn't find it.

Related work items: #5728