Merged PR 11245: Add support for searching item references from the Items smart search
Supporting searching the items by Reference No. from the Item Reference table will be useful for all of the customers and because of that we are adding this.

Also since we need to check data in another table to find the items and then get the result back we don't support the * character at the beginning of the search while searching Reference No. since this will have a huge performance impact on the result set.

Related work items: #5806