Merged PR 11180: Report to check fields not handled by Replication
It checks the API page fields using the $metadata of web services and compares the fields from the response xml with the table fields.
Then prints in the layout any missing fields (which are present in the table, but not in the API $metadata response).

Fixed also found missing fields in existing core APIs.

Related work items: #5774