Merged PR 11253: Adding Inventory to NPR Item Variants page
Many customers need to have information about Inventory for each Item Variant and having that information on the factbox isn't good enough for most of them.
"NPR Item Variants Lookup" page needs to be blazingly fast since it is opened while doing a sale. Because of that, we added inventory information on the "NPR Item Variants" page and then created an action for opening this page from the POS.

The reason, why we created a new action for this instead of just using action for RUNPAGE or RUNOBJECT, is that we need to filter Inventory based on the Location, and there is no way of sending that information to those above-mentioned actions.

Related work items: #5750