Merged PR 11428: Browser stored creden. workar. in POS View Profile
Hi @<Mark Stewart Pedersen>

There is interesting difference between BC 17 and 19 with changing value in field which is Primary Key:

In 17, you are able to change from blank to something, but not from something to blank. (Which is absolutely OK)
But in 19, you are not able to change "empty" at all. Neither from empty to something. (Which is I would say crazy but never mind...)

Problem in case 508803 started from another issue which we have for some time:
Internet Browsers, no matter if it is Microsoft Edge, Google Chrome, Firefox or Opera, will automatically fill some fields with username or email field if credentials are saved in that browser!
On Magento Setup quite often browser is changing username and password fields for API.
In this case it is filling "POS Theme Code" field. (this is even more crazy!)

To make this story short, with switch to BC 19 I guess that this will not be only page where we will have stupid issue like here.
When page is opened, browser will automatically fill some filed ("POS Theme Code" for example):

![image.png](https://dev.azure.com/navipartner/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/11428/attachments/image.png)

Of course table relation will fail so you will have an error and need to fix it before you make next click. And when you change it, "Code" filed is empty and because it is Primary Key and this is BC19, you are not able to change it. 🤣

Quick workaround for user is to delete it, start InPrivate/Incognito browsing and then create new record. Because InPrivate, browser will not fill anything.

Permanent workaround is code change from this PR:

![image (2).png](https://dev.azure.com/navipartner/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/11428/attachments/image%20%282%29.png)

But as I said... it can repeat somewhere else, on some other Page.
I have googled for this if there is some property which will prevent browser to auto fill the filed, but I haven't find it. There is a setting in the browser, but we can not configure user's browsers...

I have added you here to be informed. If someone report same issue, this can help.

Related work items: #5870