PR 11381: Fixed saved sale SystemId bug 2
This is in addition to this PR: !6819
>Previously every time a sale was saved this created a cancelled sale entry in POS entry table.
This did not really fit into the new SystemId lifecycle where the ID of a sale is only present in one table at any given moment (Active Sale, Saved Sale or POS entry).

Same issue here: when a previous date sale is resumed, what system does in the background is the following: it parks the sale (saves it as a quote), and then retrieves the parked sale.
The problem was that when system would park a sale, it would also cancel it. This led to the same SystemId be used for both parked sale and POS entry, and as a result there would be SystemId clash, when we attempted to retrieve and post the parked sale.

The solution here is the same, as it was in the PR mentioned above: I have added a system event to POS entry for when sale is parked, so we still have traceability without messing with the sales SystemId logic.

I have also added a way to handle currently existing cases, when there's already a parked POS sale and a POS entry with the same SystemId: when a parked sale is retrieved, system will check if there is a POS entry with the same SystemId, and if there is an entry, and the entry is of type "Cancelled Sale", system will just change SystemId of the POS entry.

Case 509167.
Related work items: #4406