Merged PR 11465: Job Task Dimens. being overwr. on Invoice creation
Dimensions from Job Task being overwritten on Sales Line when Invoice is created from Job Task

This looks like Microsoft bug. I don't know if they fixed it already or not (we are going to create a ticket).
And after we figure out in which version it's fixed, we will add tag in code.

Tihomir's and my opinion is this should go into NpCore because it is needed for all clients who are using Event module.

Two local functions and block of code are copied from standard codeunit 1002 "Job Create-Invoice"
I have only added to exit if "Prices Including VAT" on Customer is false because in that case, dimension are not being overwritten. And to exit if "Job Task No." is empty because in that case, there is nothing to pick dimensions from.

Case 501105

Related work items: #5876