Merged PR 11449: Improvements to payment method on SALES_DOC_EXP
**Improvements to payment method on SALES_DOC_EXP**

Previously (C/SIDE versions) we were able to keep the default payment method code inherited from the customer. [During the clean up of Payment Types](https://dev.azure.com/navipartner/NpCore/_git/NpCore/pullrequest/7417), a new parameter was added to specify the Payment Method Code directly on the POS Menu Button.

This solution was good, but unfortunately it was implemented in a way that would override any payment method code set on the sales header by default with an empty value if the button did not have a method specified. This PR introduces another parameter that enables configuration as to where the payment method code should come from.

We now offer possibility to inherit default, force a blank code or use the specified code on the POS menu button. The upgrade code will ensure that any customers that were using a specific method will not be affected by this.

Can be tested in container linked to case 509793.

Related work items: #5868