PR 11622: Cashless end of day routine
Improved the "end of day" (EoD) routine to support cashless scenarios (when there are no cash sales registered for the day).
Now system will do the auto-count routine (the one it did for payment types with value 'virtual' set for field 'include in counting') immediately when the EoD statistics page is opened without waiting for the user to go to the counting page.
Therefore users won't need to open the counting page anymore.

But system will do it only when both of the following are true:
1. there are no non-zero (with payment amount <> 0 for the day) POS payment methods with field 'include in counting' set to value 'Yes';
2. we are in 'final counting' (Z report) mode.

In other cases system will follow the usual routine we had in place before this change.

Case 464472
Related work items: #5937