Merged PR 11172: Add Responsibility Center when importing WEB Order
This is a new feature in web order import
As described in work item, request initially came from TipTop but we agreed to make it possible in NP Retail.

Why to be in NP Retail?:
 - Because we already included new feature 9 days ago for Different logos based on responsibility center,
    and same thing was 2nd part of the request from this case.
 - Because we skipped creation of separate codeunit in client extension and by that we are keeping NP core
    standard with this client and all feature updates/bug fixes
 - It is designed like that so it can support Magento web shop clients plus other webshop solutions.
 - It will not affect/stop any of current customers to import their orders.

   And here is how it will not affect: Element "responsibility_center" is optional.
   - If Doesn't exist, it will just ignore it and go as it was going till now.
   - If exist and filled in the incoming XML file, it will be passed to Sales Header.
   - If exist but not filled, then it will go to Magento Website table (based on website_code element) and
  search for it there:
       - If it is filled in, it will be passed to Sales Header.
       - If not, it will just go without it.

Also, I have added new codeunit for manual event subscribers.
I didn't wanted to extend existing one because it is made because of Salesperson inserting.
But I also didn't wanted to rename it because it was made a long time ago and placed into "POS Sales Document" folder.
Instead, I made new codeunit with general name so we can use it in the future when needed.

Related work items: #5761