PR 11817: POS Sale Resume proces: removed 2nd warning message
Removed 2nd warning, which was shown by the system during POS Sale resume process, if the sale being processed belonged to another active session.
The 2nd prompt was originally added in this PR: !10953
The problems was that a user could seize POS session from another user, working under the same user ID.
However, while trying to fix an edge case, we introduced a prompt which would confuse users in much bigger number of cases.
So in this PR we are removing the prompt and adding a more meaningful error message, when a user ends up in a non-existing (seized by another another user session) sales ticket.
From now onwards, if we get the message in telemetry, the case should be handled by support by fixing setups and educating users, so nobody uses other people user IDs.

Case 511326
Related work items: #5666