Merged PR 11844: Add SelectLatestVersion in Task list processing
Long story short we were observing that sometimes the data sent to external parties was not up to date.

Through investigation we figured out this was because the NAS service was reading a record, the user updating and the NAS accessing the record again, all within the 30 second gap that the NSTs allow for "cache drift" between each other.

This PR forces the task list processing to re-read records so we ensure that the data we send or process upon is _always_ up to date.

Related work items: #6050