Merged PR 11853: Magento Order Comments not being saved as Notes
Comments from Magento Sales Order are not being saved as Notes in BC

This issue existed before. It was reported in the same case and solved through case 465719.
But now error came again because of PR 10395

Now it's fixed and works well. Tested in case container. And also with multiple comments and one longer one:

![image.png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/11853/attachments/image.png)

Related work items: #6054