Merged PR 11838: Payment workflows refactored to next generation of workflow capabilities.
#Noteable changes

* Workflows version 3 are now declared as a function such as:
`let main = async ({ workflow, context, popup, runtime, hwc, data, parameters, captions, scope}) => {`
`  // your workflow code here`
` `
`}`
Note that the  workflow object sent as a parameter can be destructured for simpler access of the different members.

* Workflow should now be declared in separate .js file. This results in full editor language support. A minified version of the JavaScript is written to its "importing" file.
This requires that "Run On Save" extension is installed with VS-Code. (run on save will run a ps-script and check for a specific tag in the .al file and insert a minified version of the .js file) Include the following in you .al file:
//###NPR_INJECT_FROM_FILE:**_<Name of .js file, must start with POSAction and have the .js extension>_**###
`    local procedure GetActionScript(): Text`
`    begin`
`        exit(`
`//###NPR_INJECT_FROM_FILE:POSActionEFTMock.Codeunit.js###`
`' ');`
`    end;`

* The POS Action uses an Enum to register. See "NPR POS Workflow".
* The `Register()` is called by framework to register the POS Action.

#Other changes include

* POS Session is single instance. So no need to pass it around as argument to your functions.
* `Refresh()` is not required to be called by the actions and is now handled by the framework.
* `RunWorkflow()` (aka OnAction, OnAction20) exposes the more common codeunits you would normally use.

###See an example in "NPR POS Action: EFT Mock".

#Hardware Connector

To interact with local hardware, there is a new API. The hardware connector is called from the workflow JavaScript and is a service running on the local computer. Include the `hwc` class in your workflow.

###See an example in "NPR POS Action: EFT Mock".

Related work items: #6047