Merged PR 11929: Ecommerce The XPath query '/post_code' does not match any nodes.
I have found this error by accident. I have configured empty case container for E-Commerce import so customer wasn't there. It goes well (customer is being created and order imported successfully, import list record is marked as imported but error message is visible).

![image.png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/11929/attachments/image.png)

It is the same for "Ecommerce The XPath query '/country_code' does not match any nodes."
I have changed both.

P.S. In our container when it's created with base data, it may happen that we will get error:

![image (2).png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/11929/attachments/image%20%282%29.png)

But this is because in our RapidStart package we have customers with D000002, *3, *4, *5, and *6 but in Number Series for Customer last number used can be D000001 after one customer is created so it will fail for next 5 customer creation attempts (or you can go and manually change last used number)

Real fix for this should be that our RapidStart shouldn't contain same Numbers as from Customer Nos. ;)

Related work items: #6090