Merged PR 12071: POS Default view was not selected as it should
There were scenarios when Default View wasn't selected as it should, for example, if in the table exists record with the POS Unit No. Filter. This bug is first noticed on the OttoSuenson client in case 492722.
Also, in this PR the code for selecting the views is refactored to be more clear.

Related work items: #6166