Merged PR 12173: Added formatting fixes for the Change Ticket Reservation article
- I've only added stylistic and formatting fixes, but feel free to review the article just in case.
@<Miemaria Syhler> @<Lars Peter Thorup>

As for the Authorization Code Scheme and Authorization Code segments, my suggestion is to remove them from here and use information about them that Miemaria wrote to create an independent explanation article on the subject. That article can be linked inside Change Ticket Reservation.
In that way, we would reduce this break that currently exists between the steps, which would improve the usability of the article.
If this sounds reasonable to you, I'll create a case.

Related work items: #6175