Merged PR 12275: E-Commerce order import card_type not registered
E-Commerce order import card_type not registered

When importing E-Commerce order, card_type is not registered because in codeunit it was expected as xml attribute instead of element. If we take a look to XML port 6151300 "NPR NpEc Sales Order Import", it is obvious that it is element, not attribute.

Tested in container from another case. Works good.

Related work items: #6252