Merged PR 12457: POS Statistics pages
Many of our customers need to see statistics inside the POS and because of that two new POS statistics pages are introduced in this PR:
- POS Single Statistics Page - where users can easily see details about the last sale that was made. Also, they can filter it to find another single sale and see details about that one.
![image (2).png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/12457/attachments/image%20%282%29.png)
- POS Turnover Stats page - where users can see statistics for different periods: last day, last week, last month, last year, current week, current month, etc., and compare it with the same period from the previous year.
![image.png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/12457/attachments/image.png)
They could also easily change a filter for specific Base Date and POS Stores and Units.

To retrieve statistics data I used queries to gain as much as we could performance benefits.

After this PR is approved and hotfixes are created so customers have pages to see statistics, the plan is to work together with frontend team to create these 2 pages with HTML and CSS so we can achieve a very nice design of those.

Related work items: #5121